Refactor jwt_validation some functions to new router_guards.py
